web-animations/animatable/animatable-animate-effect.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animatable.animate() method effect argument test</title>
<meta name="assert" content="The effect to apply. This value is passed to the Animation constructor as the effect parameter and has the same interpretation as defined for that constructor">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animatable-animate">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var player = animatable.animate(null);
        assert_equals(player.source.effect, null, 'Node ' + index + ': Animation.effect ' +
            'should be null');
    });
}, 'Null value of effect argument of Animatable.animate() method produces ' +
    'Animation with null effect');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect = new KeyframeEffect(KEYFRAMES);
        var player = animatable.animate(effect);
        assert_equals(player.source.effect, effect, 'Node ' + index + ': Animation.effect ' +
            'should be set by the Animatable.animate() effect parameter');
    });
}, 'AnimationEffect as a value of effect argument of Animatable.animate() method ' +
    'produces Animation with specified AnimationEffect');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect = new KeyframeEffect(KEYFRAMES);
        var animation = new Animation(null, effect);
        var player = animatable.animate(animation.effect);

        assert_true(player.source.effect === effect, 'Node ' + index + ': Animation.effect ' +
            'should be set by the Animatable.animate() effect parameter');
        assert_true(player.source.effect === animation.effect, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
    });
}, 'AnimationEffect from different Animation object instance as a value ' +
    'of effect argument of Animatable.animate() produces Animation with the same ' +
    'effect instance');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect = function(time) {
            document.documentElement.currentScale = 1.0 + time * 2.0;
        };
        var player = animatable.animate(effect);
        assert_equals(player.source.effect, effect, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
    });
}, 'EffectCallback function as a value of effect argument of Animatable.animate() method ' +
    'produces Animation with specified EffectCallback');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect = function(time) {
            document.documentElement.currentScale = 1.0 + time * 2.0;
        };
        var animation = new Animation(null, effect);
        var player = animatable.animate(animation.effect);

        assert_true(player.source.effect === effect, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_true(player.source.effect === animation.effect, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
    });
}, 'EffectCallback function from different Animation object instance as a value ' +
    'of effect argument of Animatable.animate() method produces animation with the same ' +
    'effect instance');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect =  {top: '100px', width: '300px'};
        var player = animatable.animate(effect);

        assert_equals(player.source.effect.getFrames()[0].top, effect.top, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_equals(player.source.effect.getFrames()[0].width, effect.width, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_equals(player.source.effect.getFrames()[0].offset, null, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_equals(player.source.effect.getFrames()[0].easing, 'linear', 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_equals(player.source.effect.getFrames()[0].composite, null, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
    });
}, 'Single Keyframe object as a value of effect argument of Animatable.animate() ' +
    'produces Animation with the specified effect');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect =  {top: '100px', width: '300px', offset: 0.5, easing: 'ease',
            composite: 'add'};
        var player = animatable.animate(effect);

        assert_equals(player.source.effect.getFrames()[0].top, effect.top, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_equals(player.source.effect.getFrames()[0].width, effect.width, 'Node ' + index +
            ': Animation.effect should be set by the Animatable.animate() effect parameter');
        assert_equals(player.source.effect.getFrames()[0].offset, effect.offset, 'Node ' +
            index + ': Keyframe parameters set by Animatable.animate() should have ' +
            'specified values');
        assert_equals(player.source.effect.getFrames()[0].easing, effect.easing, 'Node ' +
            index + ': Keyframe parameters set by Animatable.animate() should have ' +
            'specified values');
        assert_equals(player.source.effect.getFrames()[0].composite, effect.composite, 'Node '
            + index + ': Keyframe parameters set by Animatable.animate() should have ' +
            'specified values');
    });
}, 'Single Keyframe object with not default properties as a value of effect ' +
    'argument of Animatable.animate() produces Animation with the specified effect');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect =  [];
        var player = animatable.animate(effect);
        assert_array_equals(player.source.effect.getFrames(), effect,
            'Node ' + index + ': Animation.effect should be set by Animatable.animate()');
    });
}, 'Empty sequence of Keyframe objects as a value of effect argument ' +
    'of Animatable.animate() produces Animation with empty keyframes sequence');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect = [{top: '0px', width: '10px', offset: 0},
            {top: '20px', width: '30px', offset: 1}];
        var player = animatable.animate(effect);

        assert_equals(player.source.effect.getFrames().length, 2,
            'Node ' + index + ': Animation.effect should have 2 frames set by the constructor');

        player.source.effect.getFrames().forEach(function(value, i) {
            assert_equals(value.top, effect[i].top, 'Node ' + index + ', keyframe ' + i +
                ': Animation.effect should be set by Animatable.animate()');
            assert_equals(value.width, effect[i].width, 'Node ' + index + ', keyframe ' + i +
                ': Animation.effect should be set by Animatable.animate()');
            assert_equals(value.offset, effect[i].offset, 'Node ' + index + ', keyframe ' + i +
                ': Offset of the animation effect should be set by Animatable.animate()');
            assert_equals(value.easing, 'linear', 'Node ' + index + ', keyframe ' + i +
                ': Animation effect easing should have default value');
            assert_equals(value.composite, null, 'Node ' + index + ', keyframe ' + i +
                ': Animation effect composite should have default value');
        });
    });
}, 'Sequence of Keyframe objects as a value of effect argument ' +
    'of Animatable.animate() produces Animation with the specified effect');

test(function() {
    var animatableElements = [createDiv(this), createPseudoElement(this)];
    animatableElements.forEach(function(animatable, index) {
        var effect = [{top: '0px', width: '10px', offset: 0, easing: 'ease', composite: 'add'},
            {top: '20px', width: '30px', offset: 1, easing: 'ease-in', composite: 'replace'}];
        var player = animatable.animate(effect);

        assert_equals(player.source.effect.getFrames().length, 2,
            'Node ' + i + ': Animation.effect should have 2 frames set by the constructor');

        player.source.effect.getFrames().forEach(function(value, i) {
            assert_equals(value.top, effect[i].top, 'Node ' + index + ', keyframe ' + i +
                ': Animation.effect should be set by Animatable.animate()');
            assert_equals(value.width, effect[i].width, 'Node ' + index + ', keyframe ' + i +
                ': Animation.effect should be set by Animatable.animate()');
            assert_equals(value.offset, effect[i].offset, 'Node ' + index + ', keyframe ' + i +
                ': Offset of the animation effect should be set by Animatable.animate()');
            assert_equals(value.easing, effect[i].easing, 'Node ' + index + ', keyframe ' + i +
                ': Animation effect easing should be set by Animatable.animate()');
            assert_equals(value.composite, effect[i].composite, 'Node ' + index +
                ', keyframe ' + i + ': Animation effect composite should be set by ' +
                ': Animatable.animate()');
        });
    });
}, 'Sequence of Keyframe objects, with all properties specified, as a value of effect ' +
    'argument of Animatable.animate() produces Animation with the specified effect');
</script>
</body>
